+++ resolwe/flow/utils/__init__.py
@@ -3,5 +3,10 @@
 Flow Utilities
 ==============
 
+.. automodule:: resolwe.flow.utils.purge
+   :members:
+.. automodule:: resolwe.flow.utils.test
+   :members:
+
 """
 from __future__ import absolute_import, division, print_function, unicode_literals